djangorest/blog/requirements.txt
Django
djangorestframework
python-dotenv
whitenoise